feat: split loading tree in intergrated dashboard
