+++ packages/dashboard/src/initialization/Dashboard.tsx
@@ -1,39 +1,36 @@
+import React, { StrictMode } from 'react'
+import { HashRouter } from 'react-router-dom'
 import { CssBaseline, ThemeProvider, StyledEngineProvider } from '@material-ui/core'
 import { StylesProvider } from '@material-ui/styles'
-import { MaskLightTheme, ErrorBoundary, addMaskThemeI18N, applyMaskColorVars } from '@dimensiondev/maskbook-theme'
-import { HashRouter } from 'react-router-dom'
-import { Pages } from './pages/routes'
-import { StrictMode } from 'react'
+import { MaskLightTheme, ErrorBoundary, applyMaskColorVars } from '@dimensiondev/maskbook-theme'
+import { ChainId } from '@dimensiondev/maskbook-shared'
+import { Emitter } from '@servie/events'
+
 import i18n from 'i18next'
-import { I18nextProvider, initReactI18next } from 'react-i18next'
-import { addDashboardI18N } from './locales'
+import { I18nextProvider } from 'react-i18next'
+
 import './plugins'
 import {
     startPluginDashboard,
     createInjectHooksRenderer,
     useActivatedPluginsDashboard,
 } from '@dimensiondev/mask-plugin-infra'
-import { ChainId } from '@dimensiondev/maskbook-shared'
-import { Emitter } from '@servie/events'
 
 const PluginRender = createInjectHooksRenderer(useActivatedPluginsDashboard, (x) => x.GlobalInjection)
-i18n.init({
-    resources: {},
-    keySeparator: false,
-    interpolation: { escapeValue: false },
-    fallbackLng: 'en',
-})
-i18n.use(initReactI18next)
-addMaskThemeI18N(i18n)
-addDashboardI18N(i18n)
+
 applyMaskColorVars(document.body, 'light')
-// TODO:
+// TODO: implement
 startPluginDashboard({
     enabled: { events: new Emitter(), isEnabled: () => true },
     eth: { current: () => ChainId.Mainnet, events: new Emitter() },
 })
-
-export function App() {
+/**
+ * For isolated version, its children should be <Pages /> (sync)
+ * For intergrad version, its children should be <Suspense fallback="loading..."><Pages /></Suspense>
+ *
+ * The reason is the intergrated version might not be able to setup services synchronously
+ */
+export function DashboardBase(props: React.PropsWithChildren<{}>) {
     return (
         <StrictMode>
             <I18nextProvider i18n={i18n}>
@@ -42,9 +39,7 @@ export function App() {
                         <ThemeProvider theme={MaskLightTheme}>
                             <ErrorBoundary>
                                 <CssBaseline />
-                                <HashRouter>
-                                    <Pages />
-                                </HashRouter>
+                                <HashRouter>{props.children}</HashRouter>
                                 <PluginRender />
                             </ErrorBoundary>
                         </ThemeProvider>
