+++ packages/dashboard/src/initialization/esbuild_jsx_transform.ts
@@ -1,4 +1,3 @@
-/// <reference types="react/experimental" />
 import React from 'react'
 
 // Patch for esbuild (not support JSX new transform)
