+++ packages/dashboard/src/initialization/plugins.ts
@@ -1,1 +1,2 @@
+// All enabled plugin list
 import '@dimensiondev/plugin-example'
